requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pennylane>=0.32.0
tqdm>=4.65.0
pyyaml>=6.0
pandas>=2.0.0
torch-geometric>=2.4.0
transformers>=4.30.0
datasets>=2.14.0
huggingface-hub>=0.16.0
pillow>=9.5.0
pyarrow>=12.0.0

